+++ spec/integration/shared_examples/logical_switch/delete.rb
@@ -6,17 +6,17 @@
 #
 # Unless required by applicable law or agreed to in writing, software distributed
 # under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-# CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
-# language governing permissions and limitations under the License.
+# CONDITIONS OF ANY KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations under the License.
 
-require_relative '../../api300/synergy/logical_switch'
+RSpec.shared_examples 'LogicalSwitchDeleteExample' do |context_name|
+  include_context context_name
 
-module OneviewSDK
-  module API500
-    module Synergy
-      # Logical switch resource implementation for API500 Synergy
-      class LogicalSwitch < OneviewSDK::API300::Synergy::LogicalSwitch
-      end
+  describe '#delete' do
+    it 'removes Logical Switch' do
+      item = described_class.new(current_client, name: LOG_SWI_NAME)
+      item.retrieve!
+      expect { item.delete }.to_not raise_error
     end
   end
 end