+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Unreleased Changes
+
+## Suggested release: v4.3.0
+
+#### Features supported
+This release adds support to OneView Rest API version 500 for the hardware variants C7000 and Synergy to the already existing features:
+   - Ethernet network
+   - Logical switch
+
 # v4.2.0
 
 #### New Resources: